# src/swagger.yaml
openapi: 3.0.3
info:
  title: MedVehicule API
  description: API REST de gestion de flotte de véhicules hospitaliers
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Serveur local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/photos/vehicle/{vehicleId}:
    get:
      summary: Photos d'un véhicule
      tags: [Photos]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Liste }
    post:
      summary: Ajouter une photo
      description: multipart/form-data - photo (fichier), label, bookingId
      tags: [Photos]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo: { type: string, format: binary }
                label: { type: string }
                bookingId: { type: string }
      responses:
        '201': { description: Créé }
  /api/photos/{photoId}:
    delete:
      summary: Supprimer une photo
      tags: [Photos]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: photoId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Supprimé }
...
  /api/odometer/vehicle/{vehicleId}:
    get:
      summary: Relevés compteur
      tags: [Compteur]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        '200': { description: Liste }
    post:
      summary: Enregistrer un relevé
      tags: [Compteur]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value: { type: integer, minimum: 0 }
                bookingId: { type: string }
      responses:
        '201': { description: Créé }
  /api/maintenance:
    get:
      summary: Liste des entretiens
      tags: [Entretien]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string, enum: [controle_technique, revision, reparation, autre] }
        - name: status
          in: query
          schema: { type: string, enum: [planifie, en_cours, termine, reporte] }
      responses:
        '200': { description: Liste }
    post:
      summary: Créer un entretien
      tags: [Entretien]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [vehicleId, type]
              properties:
                vehicleId: { type: string }
                type: { type: string, enum: [controle_technique, revision, reparation, autre] }
                garageId: { type: string }
                contractId: { type: string }
                scheduledAt: { type: string, format: date-time }
                odometerAt: { type: integer }
                notes: { type: string }
      responses:
        '201': { description: Créé }
  /api/maintenance/garages/list:
    get:
      summary: Liste des garages partenaires
      tags: [Entretien]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Liste }
  /api/maintenance/garages:
    post:
      summary: Créer un garage partenaire
      tags: [Entretien]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [hospitalId, name]
              properties:
                hospitalId: { type: string }
                name: { type: string }
                address: { type: string }
                contact: { type: string }
      responses:
        '201': { description: Créé }
  /api/maintenance/contracts/vehicle/{vehicleId}:
    get:
      summary: Contrats d'un véhicule
      tags: [Entretien]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Liste }
  /api/maintenance/contracts:
    post:
      summary: Créer un contrat d'entretien
      tags: [Entretien]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [vehicleId, type, startDate]
              properties:
                vehicleId: { type: string }
                type: { type: string, enum: [entretien_courant, assurance, leasing] }
                garageId: { type: string }
                startDate: { type: string, format: date-time }
                endDate: { type: string, format: date-time }
                reference: { type: string }
                notes: { type: string }
      responses:
        '201': { description: Créé }
  /api/maintenance/{id}:
    get:
      summary: Détail entretien
      tags: [Entretien]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    patch:
      summary: Modifier entretien
      tags: [Entretien]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                scheduledAt: { type: string, format: date-time }
                performedAt: { type: string, format: date-time }
                odometerAt: { type: integer }
                cost: { type: number }
                notes: { type: string }
                status: { type: string, enum: [planifie, en_cours, termine, reporte] }
      responses:
        '200': { description: OK }
...